FlaskDeployment/requirements.txt
Flask==3.0.0
torch==2.2.0
torchvision==0.17.0
segmentation-models-pytorch==0.3.3
albumentations==1.3.1
numpy>=1.26.0
tifffile==2023.7.10
Pillow==10.0.0
gunicorn==21.2.0
